+++ packages/hooks/src/use-countries.ts
@@ -1,11 +1,9 @@
-import type { ICity, ICountry, IState } from 'country-state-city';
-import { City, Country, State } from 'country-state-city';
+import type { ICountry } from 'country-state-city';
+import { Country } from 'country-state-city';
 import { useMemo } from 'react';
 
 export function useCountries(): {
   countries: ICountry[];
-  getStatesOfCountry: (countryCode: string) => IState[];
-  getCitiesOfState: (countryCode: string, stateCode: string) => ICity[];
 } {
   const countries = useMemo<ICountry[]>(
     () =>
@@ -18,20 +16,7 @@ export function useCountries(): {
     [],
   );
 
-  const getStatesOfCountry = (countryCode: string): IState[] => {
-    return State.getStatesOfCountry(countryCode);
-  };
-
-  const getCitiesOfState = (
-    countryCode: string,
-    stateCode: string,
-  ): ICity[] => {
-    return City.getCitiesOfState(countryCode, stateCode);
-  };
-
   return {
     countries,
-    getCitiesOfState,
-    getStatesOfCountry,
   };
 }